# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文章生成后端：用户/会话、文章三阶段流水线（SSE 进度）、配图、Agent 日志；开发态 VIP Mock 支付与管理端升降 VIP。",
        "title": "Wood Passage Creator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/users/list": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "分页查询用户列表（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 10，最大 100",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_app_user.UserListData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "删除用户（软删除，管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（data 一般为 null）",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/revoke-vip": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "取消用户 VIP（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/upgrade-vip": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "升级用户为 VIP（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "按 ID 查询用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "部分更新用户（仅本人）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_user.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "wood-passage-creator_internal_app_user.UpdateRequest": {
            "type": "object",
            "properties": {
                "userAvatar": {
                    "type": "string",
                    "maxLength": 1024
                },
                "userName": {
                    "type": "string",
                    "maxLength": 256,
                    "minLength": 1
                },
                "userPassword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "userProfile": {
                    "type": "string",
                    "maxLength": 512
                }
            }
        },
        "wood-passage-creator_internal_app_user.UserListData": {
            "type": "object",
            "properties": {
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wood-passage-creator_internal_module_user.User"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "wood-passage-creator_internal_module_user.User": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quota": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userAccount": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "userProfile": {
                    "type": "string"
                },
                "userRole": {
                    "$ref": "#/definitions/wood-passage-creator_internal_module_user.UserRole"
                },
                "vipTime": {
                    "type": "string"
                }
            }
        },
        "wood-passage-creator_internal_module_user.UserRole": {
            "type": "string",
            "enum": [
                "user",
                "admin",
                "vip"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin",
                "RoleVIP"
            ]
        },
        "wood-passage-creator_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie authentication. Example: session=your-session-id",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}